+++ sample-groovy/buildSrc/src/main/kotlin/Libs.kt
@@ -10,8 +10,4 @@ object Libs {
     const val com_gradle_build_scan_gradle_plugin: String =
             "com.gradle.build-scan:com.gradle.build-scan.gradle.plugin:" +
             Versions.com_gradle_build_scan_gradle_plugin
-
-    const val io_vertx_vertx_plugin_gradle_plugin: String =
-            "io.vertx.vertx-plugin:io.vertx.vertx-plugin.gradle.plugin:" +
-            Versions.io_vertx_vertx_plugin_gradle_plugin
 }