+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See https://github.com/jmfayard/buildSrcVersions/issues/55
 
 
 - Upgrade to Gradle 5.6.2
+- rejectVersionsIf { ... } from https://github.com/ben-manes/gradle-versions-plugin/issues/325
 - Order dependencies by length #65
 - Improve code quality
 - Write more and better tests